rename to jsx-one-expression-per-line
Now that we handle all nodes instead of only React elements, this name
seems more fitting.
